+++ tests/testthat/test_mutual_difference.R
@@ -12,12 +12,69 @@ test_data2 <- data.frame(
     n = c(20, 20, 20, 15, 5, 10, 10, 15)
 )
 
-test_that("mutual_difference error", {
+test_that("mutual_difference method not defined error", {
     expect_error(mutual_difference(test_data1, test_data2, "g", "u", weight = "n", method = "X"))
 })
 
-test_that("mutual_difference IPF", {
-    ret1 <- mutual_difference(test_data1, test_data2, "g", "u", weight = "n", method = "ipf",
+test_that("mutual_difference SHAPLEY", {
+    sh1 <- mutual_difference(test_data1, test_data2, "g", "u", weight = "n", method = "shapley")
+    sh2 <- mutual_difference(test_data1, test_data2, "u", "g", weight = "n", method = "shapley")
+    sh3 <- mutual_difference(test_data2, test_data1, "g", "u", weight = "n", method = "shapley")
+    sh4 <- mutual_difference(test_data2, test_data1, "u", "g", weight = "n", method = "shapley")
+
+    expect_equal(sh1[["diff", "est"]], sh1[["M2", "est"]] - sh1[["M1", "est"]])
+    expect_equal(sh1[["diff", "est"]], sum(sh1[6:8, "est"]))
+    expect_equal(sh1[["additions", "est"]], 0)
+    expect_equal(sh1[["removals", "est"]], 0)
+
+    expect_equal(sh2[["diff", "est"]], sh2[["M2", "est"]] - sh2[["M1", "est"]])
+    expect_equal(sh2[["diff", "est"]], sum(sh2[6:8, "est"]))
+    expect_equal(sh2[["additions", "est"]], 0)
+    expect_equal(sh2[["removals", "est"]], 0)
+
+    expect_equal(sh3[["diff", "est"]], sh3[["M2", "est"]] - sh3[["M1", "est"]])
+    expect_equal(sh3[["diff", "est"]], sum(sh3[6:8, "est"]))
+    expect_equal(sh3[["additions", "est"]], 0)
+    expect_equal(sh3[["removals", "est"]], 0)
+
+    expect_equal(sh4[["diff", "est"]], sh4[["M2", "est"]] - sh4[["M1", "est"]])
+    expect_equal(sh4[["diff", "est"]], sum(sh4[6:8, "est"]))
+    expect_equal(sh4[["additions", "est"]], 0)
+    expect_equal(sh4[["removals", "est"]], 0)
+
+    # symmetry by group/unit
+    expect_equal(sh1[["M1", "est"]], sh2[["M1", "est"]])
+    expect_equal(sh1[["M2", "est"]], sh2[["M2", "est"]])
+    expect_equal(sh1[["diff", "est"]], sh2[["diff", "est"]])
+    expect_equal(sh1[["group_marginal", "est"]], sh2[["unit_marginal", "est"]], tolerance = .0001)
+    expect_equal(sh1[["unit_marginal", "est"]], sh2[["group_marginal", "est"]], tolerance = .0001)
+    expect_equal(sh1[["structural", "est"]], sh2[["structural", "est"]], tolerance = .0001)
+
+    expect_equal(sh3[["M1", "est"]], sh4[["M1", "est"]])
+    expect_equal(sh3[["M2", "est"]], sh4[["M2", "est"]])
+    expect_equal(sh3[["diff", "est"]], sh4[["diff", "est"]])
+    expect_equal(sh3[["group_marginal", "est"]], sh4[["unit_marginal", "est"]], tolerance = .0001)
+    expect_equal(sh3[["unit_marginal", "est"]], sh4[["group_marginal", "est"]], tolerance = .0001)
+    expect_equal(sh3[["structural", "est"]], sh4[["structural", "est"]], tolerance = .0001)
+
+    # symmetry by data inputs
+    expect_equal(sh1[["M1", "est"]], sh3[["M2", "est"]])
+    expect_equal(sh1[["M2", "est"]], sh3[["M1", "est"]])
+    expect_equal(sh1[["diff", "est"]], -sh3[["diff", "est"]])
+    expect_equal(sh1[["group_marginal", "est"]], -sh3[["group_marginal", "est"]], tolerance = .0001)
+    expect_equal(sh1[["unit_marginal", "est"]], -sh3[["unit_marginal", "est"]], tolerance = .0001)
+    expect_equal(sh1[["structural", "est"]], -sh3[["structural", "est"]], tolerance = .0001)
+
+    expect_equal(sh2[["M1", "est"]], sh4[["M2", "est"]])
+    expect_equal(sh2[["M2", "est"]], sh4[["M1", "est"]])
+    expect_equal(sh2[["diff", "est"]], -sh4[["diff", "est"]])
+    expect_equal(sh2[["group_marginal", "est"]], -sh4[["group_marginal", "est"]], tolerance = .0001)
+    expect_equal(sh2[["unit_marginal", "est"]], -sh4[["unit_marginal", "est"]], tolerance = .0001)
+    expect_equal(sh2[["structural", "est"]], -sh4[["structural", "est"]], tolerance = .0001)
+})
+
+test_that("mutual_difference KM", {
+    ret1 <- mutual_difference(test_data1, test_data2, "g", "u", weight = "n", method = "km",
         precision = .0001)
 
     expect_equal(ret1[["diff", "est"]], ret1[["M2", "est"]] - ret1[["M1", "est"]])
@@ -27,7 +84,7 @@ test_that("mutual_difference IPF", {
     expect_equal(ret1[["removals", "est"]], 0)
 
     # other way around
-    ret <- mutual_difference(test_data1, test_data2, "u", "g", weight = "n", method = "ipf",
+    ret <- mutual_difference(test_data1, test_data2, "u", "g", weight = "n", method = "km",
         precision = .0001)
 
     expect_equal(ret[["diff", "est"]], ret[["M2", "est"]] - ret[["M1", "est"]])
@@ -57,31 +114,8 @@ test_that("mutual_difference IPF", {
 
 
 test_that("mutual_difference MRC", {
-    ret <- mutual_difference(test_data1, test_data2, "g", "u", weight = "n", method = "mrc")
-
-    expect_equal(ret[["diff", "est"]], ret[["M2", "est"]] - ret[["M1", "est"]])
-    expect_equal(ret[["diff", "est"]],
-                 sum(ret[c("unit_marginal", "group_marginal", "structural"), "est"]))
-
-    # other way around
-    ret <- mutual_difference(test_data1, test_data2, "u", "g", weight = "n", method = "mrc")
-
-    expect_equal(ret[["diff", "est"]], ret[["M2", "est"]] - ret[["M1", "est"]])
-    expect_equal(ret[["diff", "est"]],
-                 sum(ret[c("unit_marginal", "group_marginal", "structural"), "est"]))
-
-    # same as mutual_total
-    expect_equal(ret[["M1", "est"]], mutual_total(test_data1, "u", "g", weight = "n")[["M", "est"]])
-    expect_equal(ret[["M2", "est"]], mutual_total(test_data2, "u", "g", weight = "n")[["M", "est"]])
-
-    expect_equal(nrow(ret), 6)
-    expect_equal(ncol(ret), 2)
-})
-
-
-test_that("mutual_difference MRC-ADJUSTED", {
     ret <- mutual_difference(test_data1, test_data2,
-                             "g", "u", weight = "n", method = "mrc_adjusted")
+                             "g", "u", weight = "n", method = "mrc")
 
     expect_equal(ret[["diff", "est"]], ret[["M2", "est"]] - ret[["M1", "est"]])
     expect_equal(ret[["diff", "est"]],
@@ -91,7 +125,7 @@ test_that("mutual_difference MRC-ADJUSTED", {
 
     # other way around
     ret <- mutual_difference(test_data1, test_data2,
-                             "u", "g", weight = "n", method = "mrc_adjusted")
+                             "u", "g", weight = "n", method = "mrc")
 
     expect_equal(ret[["diff", "est"]], ret[["M2", "est"]] - ret[["M1", "est"]])
     expect_equal(ret[["diff", "est"]],
@@ -108,27 +142,34 @@ test_that("mutual_difference MRC-ADJUSTED", {
 
 test_that("mutual_difference SE", {
     ret <- mutual_difference(test_data1, test_data2, "g", "u", weight = "n",
-                             method = "ipf", se = TRUE, n_bootstrap = 5)
+                             method = "shapley", se = TRUE, n_bootstrap = 5)
+    expect_equal(nrow(ret), 8)
+    expect_equal(ncol(ret), 3)
+    expect_equal(all(ret[ret$est > 0, "se"] > 0), TRUE)
+
+    ret <- mutual_difference(test_data1, test_data2, "g", "u", weight = "n",
+                             method = "km", se = TRUE, n_bootstrap = 5)
     expect_equal(nrow(ret), 9)
     expect_equal(ncol(ret), 3)
     expect_equal(all(ret[ret$est > 0, "se"] > 0), TRUE)
 
     ret <- mutual_difference(test_data1, test_data2, "g", "u", weight = "n",
                              method = "mrc", se = TRUE, n_bootstrap = 5)
-    expect_equal(nrow(ret), 6)
-    expect_equal(ncol(ret), 3)
-    expect_equal(all(ret[ret$est > 0, "se"] > 0), TRUE)
-
-    ret <- mutual_difference(test_data1, test_data2, "g", "u", weight = "n",
-                             method = "mrc_adjusted", se = TRUE, n_bootstrap = 5)
     expect_equal(nrow(ret), 8)
     expect_equal(ncol(ret), 3)
     expect_equal(all(ret[ret$est > 0, "se"] > 0), TRUE)
 })
 
+
 test_that("mutual_difference log base", {
     ret <- mutual_difference(test_data1, test_data2, "g", "u",
-                             weight = "n", method = "ipf", base = 2)
+                             weight = "n", method = "shapley", base = 2)
+    expect_equal(ret[["diff", "est"]], ret[["M2", "est"]] - ret[["M1", "est"]])
+    expect_equal(nrow(ret), 8)
+    expect_equal(ncol(ret), 2)
+
+    ret <- mutual_difference(test_data1, test_data2, "g", "u",
+                             weight = "n", method = "km", base = 2)
     expect_equal(ret[["diff", "est"]], ret[["M2", "est"]] - ret[["M1", "est"]])
     expect_equal(nrow(ret), 9)
     expect_equal(ncol(ret), 2)
@@ -136,64 +177,68 @@ test_that("mutual_difference log base", {
     ret <- mutual_difference(test_data1, test_data2, "g", "u",
                              weight = "n", method = "mrc", base = 2)
     expect_equal(ret[["diff", "est"]], ret[["M2", "est"]] - ret[["M1", "est"]])
-    expect_equal(nrow(ret), 6)
-    expect_equal(ncol(ret), 2)
-
-    ret <- mutual_difference(test_data1, test_data2, "g", "u",
-                             weight = "n", method = "mrc_adjusted", base = 2)
-    expect_equal(ret[["diff", "est"]], ret[["M2", "est"]] - ret[["M1", "est"]])
     expect_equal(nrow(ret), 8)
     expect_equal(ncol(ret), 2)
 })
 
-test_data1 <- data.frame(
-    u = c(rep("a", 4), rep("b", 4), rep("c", 4)),
-    g = rep(c(1, 2, 3, 4), 3),
-    n = c(0, 30, 20, 10, 5, 10, 15, 20, 4, 3, 2, 1)
-)
-test_data2 <- data.frame(
-    u = c(rep("a", 4), rep("b", 4), rep("d", 4)),
-    g = rep(c(1, 2, 3, 4), 3),
-    n = c(40, 20, 20, 15, 5, 10, 10, 0, 1, 2, 3, 4)
-)
-
-ipf <- mutual_difference(test_data1, test_data2,
-                         "g", "u", weight = "n", method = "ipf", precision = .1)
-mrc <- mutual_difference(test_data1, test_data2,
-                         "g", "u", weight = "n", method = "mrc")
-mrc_adj <- mutual_difference(test_data1, test_data2,
-                             "g", "u", weight = "n", method = "mrc_adjusted")
-M1 <- mutual_total(test_data1, "g", "u", weight = "n")[["M", "est"]]
-M2 <- mutual_total(test_data2, "g", "u", weight = "n")[["M", "est"]]
 
 test_that("difference same as mutual_total (zero weights)", {
-    expect_equal(ipf[["M1", "est"]], M1)
-    expect_equal(ipf[["M2", "est"]], M2)
+    # test with zero weights
+    test_data1 <- data.frame(
+        u = c(rep("a", 4), rep("b", 4), rep("c", 4)),
+        g = rep(c(1, 2, 3, 4), 3),
+        n = c(0, 30, 20, 10, 5, 10, 15, 20, 4, 3, 2, 1)
+    )
+    test_data2 <- data.frame(
+        u = c(rep("a", 4), rep("b", 4), rep("d", 4)),
+        g = rep(c(1, 2, 3, 4), 3),
+        n = c(40, 20, 20, 15, 5, 10, 10, 0, 1, 2, 3, 4)
+    )
+
+    shapley <- mutual_difference(test_data1, test_data2,
+                             "g", "u", weight = "n", method = "shapley", precision = .1)
+    km <- mutual_difference(test_data1, test_data2,
+                             "g", "u", weight = "n", method = "km", precision = .1)
+    mrc <- mutual_difference(test_data1, test_data2,
+                                 "g", "u", weight = "n", method = "mrc")
+    M1 <- mutual_total(test_data1, "g", "u", weight = "n")[["M", "est"]]
+    M2 <- mutual_total(test_data2, "g", "u", weight = "n")[["M", "est"]]
+
+    expect_equal(shapley[["M1", "est"]], M1)
+    expect_equal(shapley[["M2", "est"]], M2)
+    expect_equal(km[["M1", "est"]], M1)
+    expect_equal(km[["M2", "est"]], M2)
     expect_equal(mrc[["M1", "est"]], M1)
     expect_equal(mrc[["M2", "est"]], M2)
-    expect_equal(mrc_adj[["M1", "est"]], M1)
-    expect_equal(mrc_adj[["M2", "est"]], M2)
 })
 
-test_that("forward_only", {
-    f1 <- mutual_difference(test_data1, test_data2, "g", "u",
-                            weight = "n", method = "ipf", forward_only = TRUE)
-    f2 <- mutual_difference(test_data2, test_data1, "g", "u",
-                            weight = "n", method = "ipf", forward_only = TRUE)
-    both <- mutual_difference(test_data1, test_data2, "g", "u",
-                            weight = "n", method = "ipf", forward_only = FALSE)
 
-    expect_equal(f1[["M1", "est"]], f2[["M2", "est"]], both[["M1", "est"]])
-    expect_equal(f1[["M2", "est"]], f2[["M1", "est"]], both[["M2", "est"]])
-    expect_equal(f1[["diff", "est"]], -f2[["diff", "est"]], both[["diff", "est"]])
-    expect_equal(f1[["additions", "est"]], -f2[["removals", "est"]], both[["additions", "est"]])
-    expect_equal(f1[["removals", "est"]], -f2[["additions", "est"]], both[["removals", "est"]])
-    expect_equal((f1[["unit_marginal", "est"]] + -f2[["unit_marginal", "est"]]) / 2,
-        both[["unit_marginal", "est"]])
-    expect_equal((f1[["group_marginal", "est"]] + -f2[["group_marginal", "est"]]) / 2,
-        both[["group_marginal", "est"]])
-    expect_equal((f1[["interaction", "est"]] + -f2[["interaction", "est"]]) / 2,
-        both[["interaction", "est"]])
-    expect_equal((f1[["structural", "est"]] + -f2[["structural", "est"]]) / 2,
-        both[["structural", "est"]])
+test_that("correctly identifies marginal/structural changes", {
+    test_data <- data.frame(
+        u = c(rep("a", 2), rep("b", 2), rep("c", 2)),
+        g = rep(c(1, 2), 3),
+        n = c(25, 25, 28, 2, 2, 18)
+    )
+    test_data_cols <- data.frame(
+        u = c(rep("a", 2), rep("b", 2), rep("c", 2)),
+        g = rep(c(1, 2), 3),
+        n = c(25, 25, 28, 2, 2 * 2, 18 * 2)
+    )
+    test_data_rows <- data.frame(
+        u = c(rep("a", 2), rep("b", 2), rep("c", 2)),
+        g = rep(c(1, 2), 3),
+        n = c(25, 25 * 2, 28, 2 * 2, 2, 18 * 2)
+    )
+
+    shapley_cols <- mutual_difference(test_data, test_data_cols,
+                             "g", "u", weight = "n", method = "shapley")
+    expect_equal(max(abs(shapley_cols[6:8, ][["est"]])),
+                 abs(shapley_cols[["unit_marginal", "est"]]))
+    expect_equal(shapley_cols[["structural", "est"]], 0, tolerance = .001)
+
+    shapley_rows <- mutual_difference(test_data, test_data_rows,
+                             "g", "u", weight = "n", method = "shapley")
+    expect_equal(max(abs(shapley_rows[6:8, ][["est"]])),
+                 abs(shapley_rows[["unit_marginal", "est"]]))
+    expect_equal(shapley_rows[["structural", "est"]], 0, tolerance = .001)
 })